// src/types/scholarship.ts
export type Scholarship = {
    id: string;
    title: string;
    organization: string;
    eligibility?: string;
    last_date?: string;
    apply_link?: string;
    state?: string;
    qualification?: string;
    stream?: string;
    amount?: string;
    apply_start_date?: string;
    notification_url?: string;
    detail_content?: string;
    is_active: boolean;
    created_at: string;
};